magic number et version à 1 byte

+++ server/src/ch/realmtech/server/serialize/chest/ChestSerializerV1.java
@@ -68,7 +68,7 @@ public class ChestSerializerV1 implements Serializer<Integer, Integer> {
     }
 
     @Override
-    public int getVersion() {
+    public byte getVersion() {
         return 1;
     }
 }
